+++ GardenSimulation/pom.xml
@@ -124,7 +124,7 @@
             <!-- Default configuration for running with: mvn clean javafx:run -->
             <id>default-cli</id>
             <configuration>
-              <mainClass>com.example.gardensimulation/com.example.gardensimulation.HelloApplication</mainClass>
+              <mainClass>com.example.gardensimulation/com.example.gardensimulation.MainApplication</mainClass>
               <launcher>app</launcher>
               <jlinkZipName>app</jlinkZipName>
               <jlinkImageName>app</jlinkImageName>
