--- GardenSimulation/src/main/java/com/example/gardensimulation/HelloController.java
@@ -1,14 +1,0 @@
-package com.example.gardensimulation;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HelloController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-}
