+++ GardenSimulation/src/main/resources/com/example/gardensimulation/hello-view.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.*?>
 
 <?import javafx.scene.image.ImageView?>
-<VBox alignment="CENTER" prefHeight="386.0" prefWidth="363.0" spacing="20.0" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.gardensimulation.HelloController">
+<VBox alignment="CENTER" prefHeight="386.0" prefWidth="363.0" spacing="20.0" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.gardensimulation.ViewController">
   <padding>
     <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
   </padding>
